+++ src/pine/ast/main.clj
@@ -28,20 +28,21 @@
             :order           []           ;; e.g. [{ :alias "u" :column "name" :direction "DESC" }]
 
             ;; state
-            :operation      {:type  nil
-                             :value nil} ;; [ ] 1. For post-handle. e.g. set hints if operation is table.
-                                         ;; [ ] 2. For backwards compat with version < 0.5.
-                                         ;;        If op is :table, then the context  in the api handler has one less table
+            :index           0
+            :operation       {:type  nil
+                              :value nil} ;; [ ] 1. For post-handle. e.g. set hints if operation is table.
+                                          ;; [ ] 2. For backwards compat with version < 0.5.
+                                          ;;        If op is :table, then the context  in the api handler has one less table
 
-            :current        nil          ;; alias of the current table
-            :context        nil          ;; alias of the table in context
+            :current        nil           ;; alias of the current table
+            :context        nil           ;; alias of the table in context
 
             :table-count    0
             :pending-count  0
 
             ;; post
             ;; - hints
-            :hints          {:table [] :select []}})
+            :hints          {:table [] :select [] :order []}})
 
 (defn pre-handle [state connection-id ops-count]
   (-> state
@@ -58,6 +59,7 @@
     :where (where/handle state value)
     :from (from/handle state value)
     :order (order/handle state value)
+    :order-partial (order/handle state value)
     :count (pine-count/handle state value)
     :delete-action (delete-action/handle state value)
     ;; No operations
@@ -65,10 +67,14 @@
     (update state :errors conj [type "Unknown operation type in parse tree"])))
 
 (defn handle-ops [state ops]
-  (reduce (fn [s o] (-> s
-                        (handle-op o)
-                        (update :pending-count dec)
-                        (assoc :operation o))) state ops))
+  (reduce (fn [s [i o]]
+            (-> s
+                (assoc :index i)
+                (handle-op o)  ; Pass the index and operation
+                (update :pending-count dec)
+                (assoc :operation o)))
+          state
+          (map-indexed vector ops)))  ; Pair each operation with its index
 
 (defn post-handle [state]
   (-> state