+++ src/pine/ast/order.clj
@@ -1,7 +1,11 @@
 (ns pine.ast.order)
 
 (defn handle [state value]
-  (let [current (state :current)
-        columns (map #(assoc %1 :alias current) value)]
+  (let [i       (state :index)
+        current (state :current)
+        columns (map #(-> %1
+                          (assoc :alias (or (:alias %1) current))
+                          (assoc :index i))
+                     value)]
     (-> state
         (update :order into columns))))