+++ src/pine/ast/select.clj
@@ -1,7 +1,11 @@
 (ns pine.ast.select)
 
 (defn handle [state value]
-  (let [current (state :current)
-        columns (map #(if (:alias %1) %1 (assoc %1 :alias current)) value)]
+  (let [i       (state :index)
+        current (state :current)
+        columns (map #(-> %1
+                          (assoc :alias (or (:alias %1) current))
+                          (assoc :index i))
+                     value)]
     (-> state
         (update :columns into columns))))