+++ test/pine/hints_test.clj
@@ -85,5 +85,9 @@
     (is (= [{:column "id" :alias "c_0"}]     (->  "company    | s:"                 gen :select)))
     (is (= [{:column "id" :alias "c_0"}]     (->  "x.company  | s:"                 gen :select)))
     (is (= ["reports_to"  "company_id" "id"] (->> "y.employee | s:"                 gen :select (map :column))))
-    (is (= ["reports_to"]                    (->> "y.employee | s: id, company_id," gen :select (map :column))))))
+    (is (= ["reports_to"]                    (->> "y.employee | s: id, company_id," gen :select (map :column)))))
 
+  (testing "Generate `order-partial` hints"
+    (is (= [{:column "id" :alias "c_0"}]     (->  "company | o:"         gen :order)))
+    (is (= []                                (->  "company | o: id," gen :order)))
+    (is (= [{:column "id" :alias "c_0"}]     (->  "company | s: id | o:" gen :order)))))